widget: fix data_updated hook
In all versions prior to 0.2, `data_updated()` only updated the data of a
already rendered chart, and not its config. Since 0.2 the whole chart state
gets updated, which seems not to be supported by Chart.js. After the update
the whole chart is empty.
This issue was reported by @markusgritsch.

This patch fixes this issue by restoring the old behavior.

Signed-off-by: Florian Scherf <mail@florianscherf.de>